Added UnityWebRTCServerPlugin project
This plugin associates Unity's output frame buffer with the Toolkit3D and provides callbacks to Unity for messages sent through the data channel.  

Note that the debugger will attempt to load Unity development environment and paths to the installed version of Unity may need to be updated.  Also, be sure to place the nvEncConfig and webrtcConfig json files in the editor's executable directory.