Added abstraction layer for handling user events and applying to camera state.
Added a base class DataChannelHandler and two derived classes DefaultDataChannelHandler and JsonDataChannelHandler to provide an abstraction layer to allow alternate message types, handling of client events, and camera navigation state.

The Default handler applies mouse and keyboard events to a local camera state and send text messages through the data channel to the server app.

The Json handler applies mouse and keyboard events to a local camera navigator and forwards the camera state as a json message.  It also forwards keyboard and mouse events to the server as json messages.